fix: sanitize XML 1.0-invalid chars from translation export (XLIFF, Android, ResX)
Same root cause as the TMX export crash: a user-pasted vertical tab (0x0B,
common from Word/Excel copy-paste) anywhere in a translation crashes the
whole project export. JDK's default Transformer refuses codepoints outside
the XML 1.0 Char production with TransformerException, mirroring Woodstox's
WstxIOException on the StAX side.

Production Sentry TOLGEE-BACKEND-3E3 — POST /v2/projects/29392/export, last
seen minutes ago.

Extract the sanitiser from TmxExporter into a shared io.tolgee.util helper
and apply at every user-content boundary that lands in an XML document:
- DOM Element.attr (covers any attribute write)
- DOM Element.appendXmlOrText (covers XLIFF, Android XML resources)
- XliffFileWriter.appendXmlIfEnabledOrText (xml-content disabled branch
  bypasses appendXmlOrText)
- TextToXmlResourcesConvertor.string (Android/Compose convertor entry —
  sanitised once at construction so every downstream path stays safe)
- ResxWriter (textContent assignments for value + comment)
